backend/python-email-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
jinja2==3.1.2
python-multipart==0.0.6
supabase==2.0.2
python-dotenv==1.0.0
sendgrid==6.10.0
pydantic==2.5.0
apscheduler==3.10.4 